+++ R/provider-anthropic.R
@@ -55,7 +55,7 @@ NULL
 #'
 #' @inheritParams chat_openai
 #' @inherit chat_openai return
-#' @param model `r param_model("claude-sonnet-4-20250514", "anthropic")`
+#' @param model `r param_model("claude-sonnet-4-5-20250929", "anthropic")`
 #' @param api_key `r lifecycle::badge("deprecated")` Use `credentials` instead.
 #' @param credentials `r api_key_param("ANTHROPIC_API_KEY")`
 #' @param cache How long to cache inputs? Defaults to "5m" (five minutes).
@@ -88,7 +88,7 @@ chat_anthropic <- function(
 ) {
   echo <- check_echo(echo)
 
-  model <- set_default(model, "claude-sonnet-4-20250514")
+  model <- set_default(model, "claude-sonnet-4-5-20250929")
   cache <- arg_match(cache)
 
   credentials <- as_credentials(
@@ -119,7 +119,7 @@ chat_claude <- chat_anthropic
 
 chat_anthropic_test <- function(
   ...,
-  model = "claude-3-5-sonnet-latest",
+  model = "claude-sonnet-4-5-20250929",
   params = NULL,
   echo = "none"
 ) {